split main class into it's own file
